Split resolver metadata behavior ref restoration tests
